NEW: script to transform the EvaluateSegmentation output
The organize_EvaluateSegmentation script is focused on tranforming the output data informed in segmentation comparison to a simple CSV table.
Run the segmentation comparison in command line, but writting the screen output in a text file.
After all the subjects been processed, run the script with the list of txt files.